Fix-bug en ValueAnimator
La clase ValueAnimator utilizaba un long para representar el tiempo
actual, esto causaba que la division sea entera y siempre resulte
en 0. Ahora se utiliza doubles para obtener la maxima precision

+++ Proyecto/core/src/edu/uns/galaxian/util/animator/BezierAnimator.java
@@ -1,9 +1,9 @@
 package edu.uns.galaxian.util.animator;
 
 public class BezierAnimator<T extends Number> extends ValueAnimator<T>{
-    protected long getValorActual() {
-        long tiempoActual = getTiempoActual();
+    protected double getValorActual() {
+        double tiempoActual = getTiempoActual();
         double factorValor = Math.pow(tiempoActual, 2) * (3.0d - 2.0d * tiempoActual);
-        return valorInicial.longValue() + (valorFinal.longValue() - valorInicial.longValue()) * (long) factorValor;
+        return valorInicial.doubleValue() + (valorFinal.doubleValue() - valorInicial.doubleValue()) * factorValor;
     }
 }
